Fixed Button Issue
Fixed an issue where Buttons would not work after loading a new level.

The issue was with the interaction between referenced Objects in the Unity Editor and Singleton Objects. All of the buttons were referencing the exact copy of the LevelManager script in the scene, but since LevelManager is being brought into the new scene from previous scenes, the LevelManager in that scene would destroy itself, which would leave the Object references in all of the buttons null.

This script looks for the LevelManager in the Hierarchy, then applies the expected OnClick function to the button.

If any other buttons are added to UI elements that require a reference to a Singleton Object brought in from other Scenes, this script NEEDS TO BE ATTACHED AND MUST HANDLE BUTTON CLICKS